share button

+++ NoteWriter/NotePage.xaml.cs
@@ -71,11 +71,11 @@ public partial class NotePage : ContentPage
 
 	private async void CameraButton_Clicked(object sender, EventArgs e)
 	{
-        var options = new MediaPickerOptions
-        {
-            Title = "Take a Photo"
-        };
-        var fileResult = await MediaPicker.CapturePhotoAsync(options);
+		var options = new MediaPickerOptions
+		{
+			Title = "Take a Photo"
+		};
+		var fileResult = await MediaPicker.CapturePhotoAsync(options);
 		if (fileResult != null)
 		{
 			var localPath = Path.Combine(FileSystem.Current.AppDataDirectory, fileResult.FileName);
@@ -87,14 +87,22 @@ public partial class NotePage : ContentPage
 
 	private void DeleteButton_Clicked(object sender, EventArgs e)
 	{
-        if (sender is Button deleteButton)
-        {
-            var selectedImage = deleteButton.CommandParameter as string;
-            if (!string.IsNullOrEmpty(selectedImage))
-            {
-                note.ImagePaths.Remove(selectedImage);
-                File.Delete(selectedImage);
-            }
-        }
-    }
+		if (sender is ImageButton deleteButton
+			&& deleteButton.CommandParameter is string selectedImage
+			&& !string.IsNullOrEmpty(selectedImage))
+		{
+			note.ImagePaths.Remove(selectedImage);
+			File.Delete(selectedImage);
+		}
+	}
+
+	private void ShareButton_Clicked(object sender, EventArgs e)
+	{
+		if (sender is ImageButton shareButton
+			&& shareButton.CommandParameter is string selectedImage
+			&& !string.IsNullOrEmpty(selectedImage))
+		{
+
+		}
+	}
 }
